Move filter logic to prevent import loop
